GOBMCN2-100 - Parameter INVENTORY_FILE should be mandatory in   cleanup-oracle.sh and apply-patch.sh
Scripts cleanup-oracle.sh and apply-patch.sh need some way of
knowing "where" they're supposed to run.  This is done via the
${INVENTORY_FILE} variable or the --inventory-file command
line argument.

As it currently stands, both scripts fail if neither is provided
as the parameter is optional.

Consequently this change simply makes either the parameter
mandatory and instead of failing, provides the appopriate error
message if the parameter value is not provided (enhanced user
experience).

Change-Id: Ib730b4dacfd65b23cc81831c629c9ba49c101189