GOBMCN2-110 - Bug - Error in "Get symlinks for devices" when condition
Issue identified when testing using the command line arguments:

   --ora-disk-mgmt udev

This caused step to fail as argument was not in upper case and task
"when" condition was comparing on an upper case value.  However,
value is captured in an Ansible variable and concurrently translated
to lower case.  Hence fix is to compare using the Ansible variable
and with lower case instead of re-looking up environment variables.

Change-Id: I6421b367ca639d072736ff3d1d8db78f34a42079