GOBMCN2-97 - Backlog Related - Remove parameter INSTANCE_ANSIBLE_HOSTGROUP_NAME
**********

High-level summary of changes:

  1) Change variable INSTANCE_ANSIBLE_HOSTGROUP_NAME from a command
     line argument and associated variable to a constant.

  2) Remove word "ansible" from variable's INSTANCE_ANSIBLE_HOSTNAME
     and INSTANCE_ANSIBLE_HOSTGROUP_NAME for consistancy with
     other variables related to the target instances.

Additional details and explanations of each change below.

**********

CHANGE 1: Change hostgroup variable to a constant

Small parameter adjustments based on questions raised by Jinping
during his review of the user-guide.  Specifically Jinping inquired
on the purpose of the "--instance-ansible-hostgroup-name" parameter.

The backgound is that the "hostgroup" is a legacy parameter from
initial 2019 development on VirtualBox where VMs (or hosts) can be
grouped for logical organization.  It is also used to group
targets in the Ansible inventory file.

Consequently Jinping's point is a very good one - this parameter is
not necessary for the current toolkit and it's usage for BM or even
in GCE.  However, changing the parameter and associated variable
to a constant makes the most sense as it's used for file validation
and possibly (edge case requirement) could have another use in the
future if the toolkit was expanded to support multiple clusters.

**********

CHANGE 2: Remove "ansible" from user facing variables and arguments

Other parameters and variables related to the installation target
do not include the word "ansible".  Examples: "--instance-ip-addr",
"--instance-ssh-user" and "--instance-ssh-key".

Including the word "ansible" adds to confusion for the end user as
it may imply the machine where Ansible is run when in fact the
variables refer to the target instances and not the control node
where Ansible does actually run.

Therefore, renaming for consistency with other parameters and to
remove confusion.  May at a later time discuss changing all such
parameters/variables to include the word "target" for additional
clarity.  Would handle in a separate change if discussed and
approved.

**********

Change-Id: I83c3aab6f63435fc05da1969962bc25d9212810d