Add command-line argument for swap disk
The Oracle installer requires swap to exist, so it can be convenient to
create swap as part of installation.  Formerly this was done only on GCE
VMs, and had a hard-coded default of /dev/sdg.  Here instead, we remove
the default (which may not match the install location), and permit swap
creation on any platform.  This parameter is optional, and, if not
specified, swap will not be created at all.