Update docs to work on local Jekyll
Making a few updates to permit docs to be rendered using a local Jekyll
install, allowing changes to be tested pre-commit.

There are a few components here:
- Adding yaml frontmatter --- to markdown files to tell Jekyll to
process them.  This had the side effect of adding a small table in
GitHub's view.
- Changing formatter from GFM to kramdown, which is available in vanilla
Jekyll
- Setting up an explicit default layout
- Pulling the theme directly from the upstream on build, rather than
using GitHub's built-in version

The URL for contributing and code-of-conduct pages will change to have a
.html extension and be processed when viewed on GitHub pages.