Move Jinja to the top of the backup scripts
To make it easier to test and use the script without a Jinja2 parser,
here we move all the Jinja2 code to the top declarations where they can
be easily substituted.

(And yes, an even better future state would be to remove the Jinja2
entirely and rely on command-line arguments.

Sample output (internal):
https://paste.googleplex.com/6534901617655808?raw