Select latest avilable patch
When doing a new RAC installation via rac-gi-install.yml, the patch level is set to the latest available unless oracle_rel is "base " (effectively skipping patching), or if no patches exist at all.  

The logic for patching is simpler, as our install is already at the base release.  So we explicitly fail when no matching patch could be found.

This logic only runs if oracle_rel is set to latest (the default value).  Otherwise we attempt to install the release the user specified.