Make release logic consistent across playbooks
While working on PR #81 I discovered that we have two different methods
of determine the latest release oracle_rel, and that the one in the
non-RAC installer does a lexical comparison of versions, resulting in
incorrect results (considering version 18.10 to be lower than 18.9 for
example).

This change applies the corrected logic from the RAC playbooks (which
uses an extended loop with a version compare rather than a sort), and
also removes json_query and its associated overhead and external
dependencies, in favor of pure Jinja2 selects and maps.