Simplify and speed up gcscopy
The gcscopy.yml tasks use the gsutil cat command on the control node,
and stream it to the BMS host's software library.  This mulit-stage
process avoids a dependency on Internet access or gcloud authentication
on the BMS host iteslf, and instead uses the pre-configuraed state of a
cloud VM on the control node.  Additionally, gcscopy caches patch fiels
on the BMS host to avoid expensive re-downloads when not necessary.

To protect against gsutil cp failures and partially-written files,
"unzip -t" is used before and after to validate the archive.  This is a
slow and expensive process requiring reading multiple gigabytes of data.

In thie change, we replace the "unzip -t" with an "unzip -l", listing
the archive contents.  As zip directories are at the end of the file, it
provides good detection of partially-copied files with much reduced
overhead.

At the same time, we simplify the code by removing the ls step (which is
covered by the unzip), add changed_when logic that provides user
feedback when a re-download was required, and avoid the post-download
check, as we already get good coverage by ansible's implicit check or
return codes combined with the set -o pipefail inthe shell, which
captures errors in either the gsutil or cat commands; we do this by
looiing for the lack of unzip -l's headers, which would have indicated
an existing file was validated.

Unnecessary [[ $? -eq 0 ]] is removed from shell output, as per
https://github.com/koalaman/shellcheck/wiki/SC2181 recommendations.

Sample output:
https://gist.github.com/mfielding/9e6e33d051764e314017006e1fd20557