Default disk mount options to nofail
The toolit creates filesystem mounts specified in the
`--ora-data-mounts` parameter to `install-oracle.sh`, to store
contents like the ORACLE_HOME or diag dest.  A filesystem is
created here and added to /etc/fstab to mount on startup.  If
the filesystem is somehow not mountable, systemd prevents the
system from booting normally and enters emergency mode,
resulting in a hung system if no interactive console is
available.

This change adds a default "nofail" mount option, which
prevents the failure status from being propagated to systemd,
allowing the system to still boot normally, and the underlying
issue to be fixed.

Sample output after install:

```
$ grep u01 /etc/fstab
/dev/sdb1 /u01 xfs nofail 0 0
```