Create tools/ and gen_patch_metadata.py
Here we create a new top-level directory, tools/, intended for tooling
that's useful for bms-toolkit but not strictly part of the toolkit
itself.

As the first tool, `gen_patch_metadata.py` retrieves patches from My
Oracle Support and generates `gi_patches` and `rdbms_patches` metadata
in a form the toolkit can consume.