Add assertions and retries to gen_patch_metadata
Using `gen_patch_metadata.py` for this round of Oracle patches, I wasted
some time and debugging effort by running using the wrong patch type.
Adding an assertion to validate that it's actually an OJVM combo patch,
as well as assertions that we actually found the necessary patch
numbers, and some simple retries, as Oracle's patch servers seem to be
unreliable the first few days after the patch is released.