Disable avahi only if the package exists
As documented in Red Hat doc https://access.redhat.com/solutions/25463
and MOS note 1501093.1, the avahi daemon must be disabled on RAC
clusters.  But the existing disabling code fails if the avahi package
doesn't exist at all.

So here we add a "when" clause to only attempt to disable if the package
is not installed.  We do it via ansible's native builtin.packages
interface, and then using jinja2 syntax to check for a key in the
resulting dict.

A side effect of invoking builting.packages is taht it saves package
data to a variable called, plainly enough, `packages.  But the existing
toolkit code _also_ has a variable called packages, which gets ends up
getting clobbered.  To resolve this we rename `packages` to
`required_packages`.