Remove filesystem from partitioned disks
Most of our testing happens in `/dev/mapper` devices which are left unpartitioned by the toolkit.  However, for other devices, the toolkit partitions the disk.  This can create an issue on cleanup: the partition itself gets wiped but not the filesystem.  So when the partition is recreated, the filesystem reappears and can potentially cause issues on future `mkfs` runs.

This change adds an extra step, running mkfs on the partition itself, in addition to the block device.  The first run removes filesystem signatures, and the second run removes the partition table itself.

Sample run:  https://gist.github.com/mfielding/0d2d2031e243495deb05c09310ef7dff