Set the environment variable CV_ASSUME_DISTID=OEL7 before running the RDBMS and GI installers #183
This commit introduces an environment variable, 'CV_ASSUME_DISTID,' for both grid and RDBMS installation tasks in Ansible playbooks. If the Oracle version is 18.0 or lower, the variable is set to 'OEL7', ensuring compatibility during the installation process. This change enhances the installation scripts by addressing potential issues related to Oracle version handling.